--- packages/components-react/src/UnoderedList.test.tsx
@@ -1,15 +1,0 @@
-// import '@testing-library/jest-dom';
-
-// import { render } from '@testing-library/react';
-// import { Toggletip } from '.';
-
-// describe('Toggletip', () => {
-//   it('renders a visible element', () => {
-//     const { container } = render(<Toggletip />);
-
-//     const toggletip = container.querySelector(':only-child');
-
-//     expect(toggletip).toBeInTheDocument();
-//     expect(toggletip).toBeVisible();
-//   });
-// });
