+++ packages/components-react/src/UnorderedListItem.tsx
@@ -1,0 +1,1 @@
+export { UnorderedListItem, type UnorderedListItemProps } from '@utrecht/component-library-react';
